--- DietApp/Model/Utility/Direction.swift
@@ -1,11 +1,0 @@
-//
-//  Direction.swift
-//  DietApp
-//
-//  Created by 川島真之 on 2023/07/08.
-//
-
-enum Direction: Int {
-  case next
-  case previous
-}
