+++ DietApp/Utilitis/Constants/TransitionDirection.swift
@@ -1,0 +1,13 @@
+//
+//  TransitionDirection.swift
+//  DietApp
+//
+//  Created by 川島真之 on 2024/11/26.
+//
+
+//画面遷移のボタン等で使用される列挙体
+
+enum TransitionDirection: Int {
+  case next
+  case previous
+}
