+++ DietApp/Utilitis/Constants/ValidationError + ValidationResult.swift
