+++ DietApp/Utilitis/Extensions/UIColor + AppThemeColor.swift
@@ -1,5 +1,5 @@
 //
-//  UIColor+AppThemeColor.swift
+//  UIColor + AppThemeColor.swift
 //  DietApp
 //
 //  Created by 川島真之 on 2024/11/25.
