+++ DietApp/View/TopPage/PhotoModalView.swift
@@ -11,7 +11,6 @@ protocol PhotoModalViewDelegate {
   func dismiss()
 }
 
-
 class PhotoModalView: UIView {
   
   @IBOutlet weak var scrollView: UIScrollView!
